fix: update plugin download logic for oci image support (#699)
